requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
httpx==0.28.1
python-dotenv==1.2.2
langchain-groq==1.1.2
langgraph==1.1.2
slowapi==0.1.9
